Okay, here's what was going on: the CFMcnd and CFMevp were being 'reset' everytime HPSim was called, and so needed to go through the unit conversion process everytime in order for them to be used correctly. But I'd set it up so the Unit Conversion sub was called only the first time, as the rest of the variables (as far as I know!) are not reset everytime and so only need to convert once. So, I've added an "ELSE" clause to the IF (OneTime) statement that converts the CFMs everytime. :) It's currently running through the E+ code!
It doesn't look like we're getting any Q values output, though, so I need to look into that. And it also knows it's reached equilibrium, so it only runs 1 iteration everytime HPSim is called after the first cycle or two; I need to try to figure out if that'll be a problem.